refactoring the Zendesk API wrapping code
