+++ lib/zendesk_client.rb
@@ -14,20 +14,13 @@ class ZendeskClient
 
     @client.insert_callback do |env|
       logger.info env
-      if env[:body]["user"]
-        if env[:body]["id"].nil?
-          raise ZendeskError.new("Authentication Error", env)
-        end
-      end
+      
       status_401 = env[:status].to_s.start_with? "401"
       too_many_login_attempts = env[:body].to_s.start_with? "Too many failed login attempts"
-      if status_401 || too_many_login_attempts
-        raise ZendeskError.new("Authentication Error", env)
-      end
-
-      if env[:body]["error"]
-        raise ZendeskError.new("Error creating ticket", env)
-      end
+      
+      raise ZendeskError, "Authentication Error: #{env.inspect}" if status_401 || too_many_login_attempts
+      
+      raise ZendeskError, "Error creating ticket: #{env.inspect}" if env[:body]["error"]
     end
 
     @client
